requirements.txt
# requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
pydantic
pandas
matplotlib
seaborn
networkx